move boundary to separate module to avoid import definition issues
